refactor(bench): extract exec module from bench_serving
Move the generation timing harness into exec.rs: GenTimings, the BenchModel
trait, run_timed, run_scheduler_stream, SchedulerBenchModel, and the
feature-gated DeepSeekV2LiteBenchModel + its attribution/timing helpers.
Items main.rs constructs or calls (trait, models + fields, GenTimings +
fields, the dsv2 helpers) are pub(crate).

The #[cfg(test)] module stays in main.rs — it exercises both the dsv2
helpers and build_request_metrics (a runner that stays here), so the crate
root is its natural home; it reaches exec's fns via `use exec::*`. Trimmed
the now-unused thread/mpsc/SchedulerRequest/TokenEvent imports from main.rs.

Pure code motion, no behavior change. Verified: default build clean and all
11 deepseek-v2-lite unit tests pass (only the pre-existing ParallelConfig
kimi-gated unused-import warning remains).
